SLING-13287: extract TestLogger into separate class
